Rearrange a bit.
