+++ app/src/main/java/org/wikipedia/dataclient/discussiontools/DiscussionToolsInfoResponse.kt
@@ -1,0 +1,14 @@
+package org.wikipedia.dataclient.discussiontools
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+class DiscussionToolsInfoResponse {
+    @SerialName("discussiontoolspageinfo") val pageInfo: PageInfo? = null
+
+    @Serializable
+    class PageInfo(
+            @SerialName("threaditemshtml") val threads: List<ThreadItem> = emptyList()
+    )
+}
