+++ app/src/main/java/org/wikipedia/dataclient/discussiontools/ThreadItem.kt
@@ -1,0 +1,16 @@
+package org.wikipedia.dataclient.discussiontools
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+class ThreadItem(
+        val type: String = "",
+        val level: Int = 0,
+        val id: String = "",
+        val html: String = "",
+        val author: String = "",
+        val timestamp: String = "",
+        val headingLevel: Int = 0,
+        val placeholderHeading: Boolean = false,
+        val replies: List<ThreadItem> = emptyList()
+)
